+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This is similar to [Angular's CHANGELOG.md](https://github.com/angular/angular/b
 ## [unreleased]
 
 ### Changed
+- Add consistency check in the UI
 - Add files in the self-service for Servicetemplates
 - Add error message if entered namespace contains a whitespace
 - Add filtering to namespaces to filter non allowed namespaces